gifs

+++ kamo/__init__.mojo
@@ -1,4 +1,4 @@
-alias dtype = DType.float32
+alias dtype = DType.float64
 alias simd_width = 4 * simdwidthof[dtype]()
 alias epsilon = 1e-12
